Promote selected firmware warnings to errors
Keep warning insertion idempotent while enforcing missing aggregate fields, signed comparisons, and unused locals alongside constructor reordering. Resolve every project warning from both production builds, including car-only helper compilation and the target-specific CRC32 format mismatch. Preserve the produced OBD burst-settling context field while removing only its no-op unpack.

+++ src/ble_runtime.cpp
@@ -161,7 +161,7 @@ void V1BLEClient::applyDeferredRuntimeEvents(SettingsManager& settings) {
 void V1BLEClient::drainCommandAndFollowupWork() {
     // Process phone->V1 commands (up to queue size per loop to drain any backlog)
     // Each call processes one command to minimize mutex hold time during BLE writes
-    for (int i = 0; i < MAX_PHONE_CMDS_PER_LOOP; i++) {
+    for (size_t i = 0; i < MAX_PHONE_CMDS_PER_LOOP; i++) {
         if (processPhoneCommandQueue() == 0) {
             break;
         }
